mailgun送信を単一コンポーネントに切り出した ただし汎用性について考慮せず
